requirements.txt
requests~=2.26.0
discord.py
python-dotenv~=0.19.2
async-timeout
Flask~=2.2.2
pip~=22.2.2
wheel~=0.37.1
pathlib~=1.0.1
setuptools~=63.4.3
nest_asyncio